Post run enhancements

+++ Public/Remove-LocalUserAccount.ps1
@@ -1,20 +1,17 @@
 function Remove-LocalUserAccount {
     <#
     .SYNOPSIS
-    Removes a local user account.
+    Removes a specified local user account.
   
     .DESCRIPTION
-    The Remove-LocalUserAccount function removes a specified local user account.
+    The Remove-LocalUserAccount function removes a specified local user account if it exists.
   
     .PARAMETER UserName
     The name of the local user account to be removed.
   
     .EXAMPLE
-    $params = @{
-        UserName = "TempUser"
-    }
-    Remove-LocalUserAccount @params
-    Removes the local user account named TempUser.
+    Remove-LocalUserAccount -UserName "MigrationInProgress"
+    Removes the local user account named "MigrationInProgress".
     #>
   
     [CmdletBinding()]
@@ -30,9 +27,17 @@ function Remove-LocalUserAccount {
   
     Process {
         try {
-            Write-EnhancedLog -Message "Removing local user account: $UserName" -Level "INFO"
-            Remove-LocalUser -Name $UserName -ErrorAction Stop
-            Write-EnhancedLog -Message "Successfully removed local user account: $UserName" -Level "INFO"
+            # Check if the user account exists
+            $user = Get-LocalUser -Name $UserName -ErrorAction SilentlyContinue
+            
+            if ($null -eq $user) {
+                Write-EnhancedLog -Message "User account '$UserName' does not exist. No action taken." -Level "WARNING"
+            }
+            else {
+                Write-EnhancedLog -Message "Removing local user account: $UserName" -Level "INFO"
+                Remove-LocalUser -Name $UserName -ErrorAction Stop
+                Write-EnhancedLog -Message "Successfully removed local user account: $UserName" -Level "INFO"
+            }
         }
         catch {
             Write-EnhancedLog -Message "An error occurred in Remove-LocalUserAccount function: $($_.Exception.Message)" -Level "ERROR"
@@ -44,10 +49,7 @@ function Remove-LocalUserAccount {
     End {
         Write-EnhancedLog -Message "Exiting Remove-LocalUserAccount function" -Level "Notice"
     }
-  }
-  
-  # # Example usage
-  # $params = @{
-  #   UserName = "TempUser"
-  # }
-  # Remove-LocalUserAccount @params
+}
+
+# Example usage
+# Remove-LocalUserAccount -UserName "MigrationInProgress"
